feature: TC new instructions Added new instructions to the TEST CORE: TEST_1: just basic, plain test TEST_2: tests or provides termination TEST_3: IO test TEST_4: tests new core creation(same group) TEST_5: tests new core creation(different group) TEST_6: logs some information of the core TEST_7: tests the creation of a new group TEST_8: logs group details of the core TEST_9: logs system details(BASIC INFO)
feature: MerryList is now generic

With the backdraw of added unnecessary bloat for every new list declared, the list is now generic and each component may define its own as it sees fit or build a wrapper to help itself further.
This decision stems from the hard-to-understand generic list previously utilized which was more preferrable but the trade-off is worth-it in my view.

bugs: Graves, the Manager that is full of flaws

A lot of BUGS related to Graves were fixed. The very set of initialization steps were full of holes and needed to be fixed. At least 10 bugs were fixed(I couldn't keep track). The bugs mainly revolved around
the groups and core logic as well as how they were being handled by Graves. You may check out the diff to identify the specific changes.

tests: More testing done

As you can see in **test/** directory, we have new tests- tc_test_3, tc_test_4, test_5- as well as the corresponding input files in **test_results/** directory. Each test tests one of the TC instruction. The results
may be found in the **test_results/assets/** directory.