feature: RIF input file format for RBC Added the new file format for RBC to use for its input. The docs for this will be available soon.
feature: RBC's Memory in place
RBC now has a memory that is can use as needed.

Changes: RBC's binary are no longer compatible
If the binary for RBC was compiled in one endianness then it won't run on other endianness.

docs: The New 3.md doc
The new 3.md doc describes Merry's MMF file format which is extremely simple. It may be found in **docs/**